+++ examples/drop_down_table/drop_down/index.jsx
@@ -11,7 +11,7 @@ function mapOption({ breed }) {
 export function Example() {
   const [value, setValue] = useState(null);
   const ref = useRef();
-  const [style, onLayoutListBox] = useConfineListBox();
+  const [onLayoutListBox] = useConfineListBox();
 
   return (
     <>
@@ -30,7 +30,6 @@ export function Example() {
         onValue={setValue}
         onLayoutListBox={onLayoutListBox}
         options={cats}
-        renderListBox={(props) => <div style={style} {...props} />}
         columns={columns}
         mapOption={mapOption}
       />