+++ src/components/highlighters/highlight_value.jsx
@@ -6,39 +6,38 @@ function emptyHighlight(highlight) {
   return !highlight.length || (highlight.length === 1 && typeof highlight[0] === 'string');
 }
 
-export function highlightValue(
-  highlighter,
-  { children },
-  { search },
-  { value, visuallyHiddenClassName },
-  { inverse, search: _search, ...options } = {},
-) {
-  const highlighted = highlighter(children ?? '', _search ?? (search || value?.label), options);
+export function highlightValue(highlighter, { inverse, search: _search, ...options } = {}) {
+  // eslint-disable-next-line react/prop-types
+  return ({ children }, state, componentProps) => {
+    const { value, visuallyHiddenClassName } = componentProps;
+    const { search } = state;
+    const highlighted = highlighter(children ?? '', _search ?? (search || value?.label), options, state, componentProps);
 
-  if (emptyHighlight(highlighted)) {
-    return highlighted.join('');
-  }
+    if (emptyHighlight(highlighted)) {
+      return highlighted.join('');
+    }
 
-  if (isIE()) {
-    // IE ignores aria-hidden, however it also doesn't suffer from the inline element issue
-    return (
-      <Highlight inverse={inverse}>
-        {highlighted}
-      </Highlight>
-    );
-  }
-
-  // Accessible naming treats inline elements as block and adds additional white space
-  return (
-    <>
-      <span className={visuallyHiddenClassName}>
-        {highlighted.join('')}
-      </span>
-      <span aria-hidden="true">
+    if (isIE()) {
+      // IE ignores aria-hidden, however it also doesn't suffer from the inline element issue
+      return (
         <Highlight inverse={inverse}>
           {highlighted}
         </Highlight>
-      </span>
-    </>
-  );
+      );
+    }
+
+    // Accessible naming treats inline elements as block and adds additional white space
+    return (
+      <>
+        <span className={visuallyHiddenClassName}>
+          {highlighted.join('')}
+        </span>
+        <span aria-hidden="true">
+          <Highlight inverse={inverse}>
+            {highlighted}
+          </Highlight>
+        </span>
+      </>
+    );
+  };
 }