+++ src/components/highlighters/pass_through_highlight.jsx
@@ -1,6 +1,6 @@
 import { highlightValue } from './highlight_value';
 import { passThroughHighlighter } from '../../highlighters/pass_through_highlighter';
 
-export function passThroughHighlight(...args) {
-  return highlightValue(passThroughHighlighter, ...args);
+export function passThroughHighlight(options) {
+  return highlightValue(passThroughHighlighter, options);
 }