+++ src/components/highlighters/pass_through_highlight.test.jsx
@@ -3,7 +3,7 @@ import { passThroughHighlight } from './pass_through_highlight';
 
 it('returns the children as is', () => {
   const { container } = render((
-    passThroughHighlight({ children: 'bar' }, { search: 'foo' }, {}, { search: 'foe' })
+    passThroughHighlight()({ children: 'bar' }, { search: 'foo' }, {}, { search: 'foe' })
   ));
 
   expect(container).toContainHTML('<div>bar</div>');